frame_msg/requirements.txt
frame-ble==1.1.1
frame-msg==5.2.1
bleak
numpy
pillow
lz4
keyboard
opencv-python
pvspeaker
